+++ module-apps/application-messages/windows/ThreadViewWindow.cpp
@@ -148,8 +148,9 @@ namespace gui
         SMS.dbsize = threadDetails->msgCount;
 
         if (threadDetails != nullptr && threadDetails->isUnread()) {
-            threadDetails->unreadMsgCount = 0;
-            DBServiceAPI::ThreadUpdate(application, *threadDetails);
+            auto app = dynamic_cast<app::ApplicationMessages *>(application);
+            assert(app != nullptr);
+            app->markSmsThreadAsRead(threadDetails->ID);
         }
 
         LOG_DEBUG("start: %d end: %d db: %d", SMS.start, SMS.end, SMS.dbsize);