+++ changelog.md
@@ -2,8 +2,15 @@
 
 ## Current version
 
+### Added
+* `[sms]` Added option to mark conversation as read/unread 
+
+### Changed
+
 ### Fixed
 * `[phonebook]` Fix wrong text label for back action.
+
+### Other
 
 ## [0.28.1] - 2020-07-10